requirements.txt
ccxt==1.79.2
python-dotenv==0.20.0